refactor: 설정파일 분리/통합
