+++ backend/src/main/java/com/yigongil/backend/config/ActuatorConfig.java
@@ -1,0 +1,16 @@
+package com.yigongil.backend.config;
+
+import org.springframework.boot.actuate.trace.http.InMemoryHttpTraceRepository;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+
+@Profile(value = {"prod", "dev"})
+@Configuration
+public class ActuatorConfig {
+
+    @Bean
+    public InMemoryHttpTraceRepository inMemoryHttpTraceRepository() {
+        return new InMemoryHttpTraceRepository();
+    }
+}
