+++ backend/src/test/java/com/yigongil/backend/ui/MemberControllerTest.java
@@ -14,7 +14,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yigongil.backend.application.MemberService;
-import com.yigongil.backend.config.LocalWebConfig;
+import com.yigongil.backend.config.AuthConfig;
 import com.yigongil.backend.config.auth.AuthContext;
 import com.yigongil.backend.config.auth.JwtTokenProvider;
 import com.yigongil.backend.domain.member.Member;
@@ -36,7 +36,7 @@ import org.springframework.context.annotation.Import;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
-@Import(LocalWebConfig.class)
+@Import(AuthConfig.class)
 @WebMvcTest(MemberController.class)
 class MemberControllerTest {
